Cleanup in text.io
- add types
- move tests out
- extend split_and_check to allow multiple lengths
